Implement multiple packages to organize files into a by-feature style.
