+++ src/main/java/com/crosa/learning/school/SchoolDto.java
@@ -1,4 +1,4 @@
-package com.crosa.learning;
+package com.crosa.learning.school;
 
 public record SchoolDto(
         String name
