+++ src/main/java/com/crosa/learning/student/StudentController.java
@@ -1,4 +1,4 @@
-package com.crosa.learning;
+package com.crosa.learning.student;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
