+++ src/main/java/com/crosa/learning/student/StudentService.java
@@ -1,4 +1,4 @@
-package com.crosa.learning;
+package com.crosa.learning.student;
 
 import org.springframework.stereotype.Service;
